migrating to PGPv4 and update SLAC GEN4 card to same FW/SW interface as KCU1500
