--- python/lcls2_pgp_fw_lib/hardware/XilinxKcu1500/__init__.py
@@ -1,13 +1,0 @@
-#-----------------------------------------------------------------------------
-# This file is part of the LCLS2 PGP Firmware Library'. It is subject to
-# the license terms in the LICENSE.txt file found in the top-level directory
-# of this distribution and at:
-#    https://confluence.slac.stanford.edu/display/ppareg/LICENSE.html.
-# No part of the LCLS2 PGP Firmware Library', including this file, may be
-# copied, modified, propagated, or distributed except according to the terms
-# contained in the LICENSE.txt file.
-#-----------------------------------------------------------------------------
-
-from lcls2_pgp_fw_lib.hardware.XilinxKcu1500._Kcu1500TimingRx import *
-from lcls2_pgp_fw_lib.hardware.XilinxKcu1500._Kcu1500Hsio import *
-
