--- python/lcls2_pgp_fw_lib/hardware/shared/__init__.py
@@ -1,6 +1,0 @@
-
-from lcls2_pgp_fw_lib.hardware.shared._Root             import *
-from lcls2_pgp_fw_lib.hardware.shared._RogueStreams     import *
-from lcls2_pgp_fw_lib.hardware.shared._TimingRx         import *
-from lcls2_pgp_fw_lib.hardware.shared._Hsio             import *
-from lcls2_pgp_fw_lib.hardware.shared._TimingPhyMonitor import *
