+++ python/lcls2_pgp_fw_lib/shared/_RogueStreams.py
@@ -8,6 +8,17 @@
 # contained in the LICENSE.txt file.
 #-----------------------------------------------------------------------------
 
-from lcls2_pgp_fw_lib.hardware.SlacPgpCardG4._SlacPgpCardG4TimingRx import *
-from lcls2_pgp_fw_lib.hardware.SlacPgpCardG4._SlacPgpCardG4Hsio import *
+import rogue
+import pyrogue.interfaces.simulation
 
+class RogueStreams(object):
+    def __init__(self,
+                 host     = 'localhost',
+                 basePort = 7000,
+                 numLanes = 4,
+                 pgp4     = False,
+                 **kwargs):
+
+        trigIndex = 32 if pgp3 else 8
+        self.pgpStreams = [[rogue.interfaces.stream.TcpClient(host, basePort+(34*lane)+2*vc) for vc in range(4)] for lane in range(numLanes)]
+        self.pgpTriggers = [pyrogue.interfaces.simulation.SideBandSim(host, basePort+(34*lane)+trigIndex) for lane in range(numLanes)]
