+++ python/lcls2_pgp_fw_lib/shared/__init__.py
@@ -1,0 +1,6 @@
+
+from lcls2_pgp_fw_lib.shared._Root             import *
+from lcls2_pgp_fw_lib.shared._RogueStreams     import *
+from lcls2_pgp_fw_lib.shared._TimingRx         import *
+from lcls2_pgp_fw_lib.shared._Hsio             import *
+from lcls2_pgp_fw_lib.shared._TimingPhyMonitor import *
