+++ shared/ruckus.tcl
@@ -8,11 +8,11 @@ if { [VersionCheck 2018.2 ] < 0 } {
 
 # Check for submodule tagging
 if { [info exists ::env(OVERRIDE_SUBMODULE_LOCKS)] != 1 || $::env(OVERRIDE_SUBMODULE_LOCKS) == 0 } {
-   if { [SubmoduleCheck {axi-pcie-core}    {3.5.2}  ] < 0 } {exit -1}
+   if { [SubmoduleCheck {axi-pcie-core}    {3.5.3}  ] < 0 } {exit -1}
    if { [SubmoduleCheck {l2si-core}        {3.3.1}  ] < 0 } {exit -1}
    if { [SubmoduleCheck {lcls-timing-core} {3.2.4}  ] < 0 } {exit -1}
    if { [SubmoduleCheck {ruckus}           {2.9.2}  ] < 0 } {exit -1}
-   if { [SubmoduleCheck {surf}             {2.12.3} ] < 0 } {exit -1}
+   if { [SubmoduleCheck {surf}             {2.13.0} ] < 0 } {exit -1}
 } else {
    puts "\n\n*********************************************************"
    puts "OVERRIDE_SUBMODULE_LOCKS != 0"
